backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "alembic>=1.18.3",
    "configparser>=7.2.0",
    "fastapi>=0.128.1",
    "locust>=2.43.3",
    "psycopg>=3.3.2",
    "psycopg2-binary>=2.9.11",
    "pwdlib[argon2]>=0.3.0",
    "pyjwt>=2.11.0",
    "pytest>=9.0.2",
    "python-multipart>=0.0.22",
    "redis>=7.2.0",
    "sqlalchemy>=2.0.46",
    "uvicorn[standard]>=0.40.0",
]
